* workspace.xml: multiple
* .DS_Store:
* Json.cs:
* ReturnValue.cs:
* RemoteMessage.cs:
* Class1.cs:
* TransformFilter.cs:
* ReaderPoco.cs:
* PocoLoader.cs:
* TableColumn.cs:
* TransformCsvStream.cs:
* PocoEnumerator.cs:
* TransformConcatinate.cs:
* riderModule.iml:
* PocoCachedEnumerator.cs:
* dexih.connections.sql.cs:
* dexih.connections.excel.cs:
* dexih.connections.flatfile.cs:
* dexih.functions.managedTasks.cs:
* dexih.functions.tests.async.csproj:
* dexih.connections.webservice.restful.reader.cs:

+++ src/dexih.functions/RemoteMessage.cs
@@ -62,7 +62,7 @@ namespace dexih.functions
 		public KeyValuePair[] Parameters { get; set; }
         public long HubKey { get; set; }
         public string RemoteAgentId { get; set; }
-		public long? TimeOut { get; set; }
+		public int? TimeOut { get; set; }
 
 		public string GetParameter(string key)
 		{
